refactor2: package and test cases: according to spring team feedback
