+++ rocketmq-spring-boot-starter/src/main/java/org/apache/rocketmq/spring/annotation/SelectorType.java
@@ -15,20 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.rocketmq.samples.springboot.consumer;
+package org.apache.rocketmq.spring.annotation;
 
-import org.apache.rocketmq.spring.annotation.RocketMQMessageListener;
-import org.apache.rocketmq.spring.core.RocketMQListener;
-import org.springframework.stereotype.Service;
+import org.apache.rocketmq.common.filter.ExpressionType;
 
-/**
- * RocketMQMessageListener
- */
-@Service
-@RocketMQMessageListener(topic = "${spring.rocketmq.topic}", consumerGroup = "string_consumer")
-public class StringConsumer implements RocketMQListener<String> {
-    @Override
-    public void onMessage(String message) {
-        System.out.printf("------- StringConsumer received: %s \n", message);
-    }
+public enum SelectorType {
+
+    /**
+     * @see ExpressionType#TAG
+     */
+    TAG,
+
+    /**
+     * @see ExpressionType#SQL92
+     */
+    SQL92
 }
