+++ rocketmq-spring-boot-starter/src/main/java/org/apache/rocketmq/spring/config/RocketMQConfigUtils.java
@@ -15,20 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.rocketmq.samples.springboot.consumer;
+package org.apache.rocketmq.spring.config;
 
-import org.apache.rocketmq.spring.annotation.RocketMQMessageListener;
-import org.apache.rocketmq.spring.core.RocketMQListener;
-import org.springframework.stereotype.Service;
+public class RocketMQConfigUtils {
+    /**
+     * The bean name of the internally managed RocketMQ transaction annotation processor.
+     */
+    public static final String ROCKETMQ_TRANSACTION_ANNOTATION_PROCESSOR_BEAN_NAME =
+        "org.springframework.rocketmq.spring.starter.internalRocketMQTransAnnotationProcessor";
 
-/**
- * RocketMQMessageListener
- */
-@Service
-@RocketMQMessageListener(topic = "${spring.rocketmq.topic}", consumerGroup = "string_consumer")
-public class StringConsumer implements RocketMQListener<String> {
-    @Override
-    public void onMessage(String message) {
-        System.out.printf("------- StringConsumer received: %s \n", message);
-    }
+    public static final String ROCKETMQ_TRANSACTION_DEFAULT_GLOBAL_NAME =
+        "rocketmq_transaction_default_global_name";
 }
