--- rocketmq-spring-boot-starter/src/main/java/org/apache/rocketmq/spring/starter/annotation/RocketMQTransactionListener.java
@@ -1,69 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.rocketmq.spring.starter.annotation;
-
-import org.apache.rocketmq.spring.starter.config.RocketMQConfigUtils;
-import org.springframework.stereotype.Component;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * This annotation is used over a class which implements interface
- * org.apache.rocketmq.client.producer.TransactionListener. The class implements
- * two methods for process callback events after the txProducer sends a transactional message.
- * <p>Note: The annotation is used only on RocketMQ client producer side, it can not be used
- * on consumer side.
- */
-@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Component
-public @interface RocketMQTransactionListener {
-
-    /**
-     * Declare the txProducerGroup that is used to relate callback event to the listener, rocketMQTemplate must send a
-     * transactional message with the declared txProducerGroup.
-     * <p>
-     * <p>It is suggested to use the default txProducerGroup if your system only needs to define a TransactionListener class.
-     */
-    String txProducerGroup() default RocketMQConfigUtils.ROCKETMQ_TRANSACTION_DEFAULT_GLOBAL_NAME;
-
-    /**
-     * Set ExecutorService params -- corePoolSize
-     */
-    int corePoolSize() default 1;
-
-    /**
-     * Set ExecutorService params -- maximumPoolSize
-     */
-    int maximumPoolSize() default 1;
-
-    /**
-     * Set ExecutorService params -- keepAliveTime
-     */
-    long keepAliveTime() default 1000 * 60; //60ms
-
-    /**
-     * Set ExecutorService params -- blockingQueueSize
-     */
-    int blockingQueueSize() default 2000;
-}
