--- rocketmq-spring-boot-starter/src/main/java/org/apache/rocketmq/spring/starter/config/RocketMQProperties.java
@@ -1,144 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.rocketmq.spring.starter.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@SuppressWarnings("WeakerAccess")
-@ConfigurationProperties(prefix = "spring.rocketmq")
-public class RocketMQProperties {
-
-    /**
-     * The name server for rocketMQ, formats: `host:port;host:port`.
-     */
-    private String nameServer;
-
-    private Producer producer;
-
-    public String getNameServer() {
-        return nameServer;
-    }
-
-    public void setNameServer(String nameServer) {
-        this.nameServer = nameServer;
-    }
-
-    public RocketMQProperties.Producer getProducer() {
-        return producer;
-    }
-
-    public void setProducer(RocketMQProperties.Producer producer) {
-        this.producer = producer;
-    }
-
-    public static class Producer {
-
-        /**
-         * Name of producer.
-         */
-        private String group;
-
-        /**
-         * Millis of send message timeout.
-         */
-        private int sendMessageTimeout = 3000;
-
-        /**
-         * Compress message body threshold, namely, message body larger than 4k will be compressed on default.
-         */
-        private int compressMessageBodyOverHowmuch = 1024 * 4;
-
-        /**
-         * Maximum number of retry to perform internally before claiming sending failure in synchronous mode.
-         * This may potentially cause message duplication which is up to application developers to resolve.
-         */
-        private int retryTimesWhenSendFailed = 2;
-
-        /**
-         * <p> Maximum number of retry to perform internally before claiming sending failure in asynchronous mode. </p>
-         * This may potentially cause message duplication which is up to application developers to resolve.
-         */
-        private int retryTimesWhenSendAsyncFailed = 2;
-
-        /**
-         * Indicate whether to retry another broker on sending failure internally.
-         */
-        private boolean retryAnotherBrokerWhenNotStoreOk = false;
-
-        /**
-         * Maximum allowed message size in bytes.
-         */
-        private int maxMessageSize = 1024 * 1024 * 4; // 4M
-
-        public String getGroup() {
-            return group;
-        }
-
-        public void setGroup(String group) {
-            this.group = group;
-        }
-
-        public int getSendMessageTimeout() {
-            return sendMessageTimeout;
-        }
-
-        public void setSendMessageTimeout(int sendMessageTimeout) {
-            this.sendMessageTimeout = sendMessageTimeout;
-        }
-
-        public int getCompressMessageBodyOverHowmuch() {
-            return compressMessageBodyOverHowmuch;
-        }
-
-        public void setCompressMessageBodyOverHowmuch(int compressMessageBodyOverHowmuch) {
-            this.compressMessageBodyOverHowmuch = compressMessageBodyOverHowmuch;
-        }
-
-        public int getRetryTimesWhenSendFailed() {
-            return retryTimesWhenSendFailed;
-        }
-
-        public void setRetryTimesWhenSendFailed(int retryTimesWhenSendFailed) {
-            this.retryTimesWhenSendFailed = retryTimesWhenSendFailed;
-        }
-
-        public int getRetryTimesWhenSendAsyncFailed() {
-            return retryTimesWhenSendAsyncFailed;
-        }
-
-        public void setRetryTimesWhenSendAsyncFailed(int retryTimesWhenSendAsyncFailed) {
-            this.retryTimesWhenSendAsyncFailed = retryTimesWhenSendAsyncFailed;
-        }
-
-        public boolean isRetryAnotherBrokerWhenNotStoreOk() {
-            return retryAnotherBrokerWhenNotStoreOk;
-        }
-
-        public void setRetryAnotherBrokerWhenNotStoreOk(boolean retryAnotherBrokerWhenNotStoreOk) {
-            this.retryAnotherBrokerWhenNotStoreOk = retryAnotherBrokerWhenNotStoreOk;
-        }
-
-        public int getMaxMessageSize() {
-            return maxMessageSize;
-        }
-
-        public void setMaxMessageSize(int maxMessageSize) {
-            this.maxMessageSize = maxMessageSize;
-        }
-    }
-}
