--- rocketmq-spring-boot-starter/src/main/java/org/apache/rocketmq/spring/starter/config/RocketMQTransactionAnnotationProcessor.java
@@ -1,116 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.rocketmq.spring.starter.config;
-
-import org.apache.rocketmq.client.exception.MQClientException;
-import org.apache.rocketmq.spring.starter.annotation.RocketMQTransactionListener;
-import org.apache.rocketmq.spring.starter.core.RocketMQLocalTransactionListener;
-import org.springframework.aop.support.AopUtils;
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.BeanCreationException;
-import org.springframework.beans.factory.BeanFactory;
-import org.springframework.beans.factory.BeanFactoryAware;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.config.BeanExpressionResolver;
-import org.springframework.beans.factory.config.BeanPostProcessor;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.context.expression.StandardBeanExpressionResolver;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.AnnotationUtils;
-
-import java.util.Collections;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class RocketMQTransactionAnnotationProcessor
-    implements BeanPostProcessor, Ordered, BeanFactoryAware {
-    private final static Logger log = LoggerFactory.getLogger(RocketMQTransactionAnnotationProcessor.class);
-
-    private BeanFactory beanFactory;
-    private BeanExpressionResolver resolver = new StandardBeanExpressionResolver();
-    private final Set<Class<?>> nonProcessedClasses =
-        Collections.newSetFromMap(new ConcurrentHashMap<Class<?>, Boolean>(64));
-
-    @Autowired(required = false)
-    private TransactionHandlerRegistry transactionHandlerRegistry;
-
-
-    @Override
-    public void setBeanFactory(BeanFactory beanFactory) throws BeansException {
-        this.beanFactory = beanFactory;
-        if (beanFactory instanceof ConfigurableListableBeanFactory) {
-            this.resolver = ((ConfigurableListableBeanFactory) beanFactory).getBeanExpressionResolver();
-        }
-    }
-
-    @Override
-    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
-        return bean;
-    }
-
-    @Override
-    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
-        if (!this.nonProcessedClasses.contains(bean.getClass())) {
-            Class<?> targetClass = AopUtils.getTargetClass(bean);
-            RocketMQTransactionListener listener = AnnotationUtils.findAnnotation(targetClass, RocketMQTransactionListener.class);
-            this.nonProcessedClasses.add(bean.getClass());
-            if (listener == null) { // for quick search
-                log.trace("No @RocketMQTransactionListener annotations found on bean type: {}", bean.getClass());
-            } else {
-                try {
-                    processTransactionListenerAnnotation(listener, bean, beanName);
-                } catch (MQClientException e) {
-                    log.error("failed to process annotation " + listener, e);
-                    throw new BeanCreationException("failed to process annotation " + listener, e);
-                }
-            }
-        }
-
-        return bean;
-    }
-
-    private void processTransactionListenerAnnotation(RocketMQTransactionListener anno, Object bean, String beanName)
-        throws MQClientException {
-        if (transactionHandlerRegistry == null) {
-            throw new MQClientException("Bad usage of @RocketMQTransactionListener, " +
-                "the class must work with producer RocketMQTemplate", null);
-        }
-        if (!RocketMQLocalTransactionListener.class.isAssignableFrom(bean.getClass())) {
-            throw new MQClientException("Bad usage of @RocketMQTransactionListener, " +
-                "the class must implement interface org.apache.rocketmq.spring.starter.core.RocketMQLocalTransactionListener",
-                null);
-        }
-        TransactionHandler transactionHandler = new TransactionHandler();
-        transactionHandler.setBeanFactory(this.beanFactory);
-        transactionHandler.setName(anno.txProducerGroup());
-        transactionHandler.setBeanName(bean.getClass().getName());
-        transactionHandler.setListener((RocketMQLocalTransactionListener) bean);
-        transactionHandler.setCheckExecutor(anno.corePoolSize(), anno.maximumPoolSize(),
-            anno.keepAliveTime(), anno.blockingQueueSize());
-
-        transactionHandlerRegistry.registerTransactionHandler(transactionHandler);
-    }
-
-    @Override
-    public int getOrder() {
-        return LOWEST_PRECEDENCE;
-    }
-
-}
