--- rocketmq-spring-boot-starter/src/main/java/org/apache/rocketmq/spring/starter/core/MessageHelper.java
@@ -1,33 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.rocketmq.spring.starter.core;
-
-import org.apache.rocketmq.spring.starter.support.RocketMQMessageConst;
-import org.springframework.messaging.Message;
-import org.springframework.util.Assert;
-
-public class MessageHelper {
-    public static String getTransactionId(Message message) {
-        Assert.isTrue(message != null, "message is null");
-        return (String) message.getHeaders().get(RocketMQMessageConst.TRANSACTION_ID);
-    }
-
-    public static String getKeys(Message message) {
-        Assert.isTrue(message != null, "message is null");
-        return (String) message.getHeaders().get(RocketMQMessageConst.KEYS);
-    }
-}
