+++ rocketmq-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.rocketmq.spring.starter.config.RocketMQAutoConfiguration
+org.apache.rocketmq.spring.config.RocketMQAutoConfiguration
