+++ samples/rocketmq-spring-boot-starter-sample/rocketmq-produce-demo/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.rocketmq.name-server=localhost:9876
+spring.rocketmq.nameServer=localhost:9876
 spring.rocketmq.producer.group=my-group1
 spring.rocketmq.topic=string-topic
 spring.rocketmq.orderTopic=order-paid-topic
